Make GA key an environment variable
Currently our GA key was committed in the code and hardwired. This was a
bad, bad idea, so this commit fixes it by making it part of the
configuration, where it reads it from the environment variables.

Also, it passes the config object to the nunjuck templates environment
so the templates can make use of them.

Last but not least, it adds a run-tests.sh script, which holds all the
steps necessary to run the tests. The script also passes fake env
variables so we can test these too.