Fixes nasty bug with a nasty hack, will improve with a PR to d2h
Problem: It was impossible to show renamed files
Reason: d2h name renamed files this way: "{old -> new}/name" and from the file tree
we search what wrapper to show by file name, the problem is that we don't name files
the same way that d2h does, plus, the module that d2h uses is private, can't be imported
nor we can't inject how we want to name the files.

Fix: Nasty hack to get it to work and will fix properly later. Basically we copied
verbatim the library that do this namings from d2h and add it into our code to be able
to have the same logic.

A better fix would be to inject how we want the output of the attributes generated by d2h